Restructured generator to allow debugging and fix
